fixed account deletion

+++ app.py
@@ -220,20 +220,25 @@ def get_current_date():
 
 # Delete User Reccord
 @app.route('/delete/<int:id>')
+@login_required
 def delete(id):
-    user_to_delete = Users.query.get_or_404(id)
-    name = None
-    form = UserForm()
+    if id == current_user.id:
+        user_to_delete = Users.query.get_or_404(id)
+        name = None
+        form = UserForm()
 
-    try:
-        db.session.delete(user_to_delete)
-        db.session.commit()
-        flash("User Deleted Successfully!")
-        our_users = Users.query.order_by(Users.date_added)
-        return render_template("add_user.html", form=form, name=name, our_users=our_users)
-    except:
-        flash("Whoops! There was a problem deleting user, try again...")
-        return render_template("add_user.html", form=form, name=name, our_users=our_users)
+        try:
+            db.session.delete(user_to_delete)
+            db.session.commit()
+            flash("User Deleted Successfully!")
+            our_users = Users.query.order_by(Users.date_added)
+            return render_template("add_user.html", form=form, name=name, our_users=our_users)
+        except:
+            flash("Whoops! There was a problem deleting user, try again...")
+            return render_template("add_user.html", form=form, name=name, our_users=our_users)
+    else: 
+        flash("You are not authorized to delete that user!")
+        return redirect(url_for('dashboard'))
 
 
 # Update Database Record
